requirements.txt
azure-functions
azure-identity>=1.17.0
azure-ai-projects>=1.0.0b9